rename: 컴포넌트 명에 page단어 추가
