+++ src/pages/WriteMapSearchPage/WriteMapSearchPage.styles.ts
