+++ src/pages/index.ts
@@ -12,4 +12,4 @@ export * from './WriteMapPage/WriteMapPage';
 export * from './WritePage/WritePage';
 export * from './SearchResultPage/SearchResultPage';
 export * from './SearchResultMapPage/SearchResultMapPage';
-export * from './WriteMapSearch/WriteMapSearch'
+export * from './WriteMapSearchPage/WriteMapSearchPage'
